+++ Changelog.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 - Display missing lines in coverage report (#376).
+- Add "Europe Day" for Luxembourg (#377).
 
 ## v5.1.0 (2019-06-24)